Rename filter→filter-sidebar vr test
